fix div: np.float32 / int -> np.float64

+++ hw/hw2/python/needle/ops/ops_mathematic.py
@@ -135,7 +135,7 @@ class DivScalar(TensorOp):
 
     def compute(self, a):
         ### BEGIN YOUR SOLUTION
-        return array_api.divide(a, self.scalar)
+        return array_api.divide(a, self.scalar).astype(a.dtype)
         ### END YOUR SOLUTION
 
     def gradient(self, out_grad, node):
